add auth ui
